support pointer events
The old model of taking unions over ActionSpaces and then masking to a specific ActionSpace didn't make much sense for pointer events. For one thing, every game has different screen dimensions, so it would be difficult to get a union over pointer events that could be meaningfully masked. Moreover, I wanted the network to always output coordinates in [-1, 1], and then the meaning of those coordinates would depend on the particular game's screen dimensions.